Move hassle to a blueprint
This is a work-in-progress. I'm still moving things around, but it looks
like this still works.
