+++ templates/hassle_new_confirm.html
@@ -3,7 +3,7 @@
 <h2>Start new hassle</h2>
 <p>Please verify that the participating members and rooms is correct.</p>
 
-<form action="{{ url_for('new_hassle_confirm_submit') }}" method="POST">
+<form action="{{ url_for('hassle_bp.new_hassle_confirm_submit') }}" method="POST">
   <div class="table_container">
     <table>
       <thead>
@@ -41,7 +41,7 @@
     </table>
   </div>
   <button type="button"
-    onclick="location.href='{{ url_for('new_hassle_participants') }}'">
+    onclick="location.href='{{ url_for('hassle_bp.new_hassle_participants') }}'">
     Edit participants/rooms
   </button>
   <button type="submit">Confirm</button>
