+++ templates/hassle_new_participants.html
@@ -3,7 +3,7 @@
 <h2>Start new hassle</h2>
 <p>Select everyone who will be participating in this room hassle.</p>
 
-<form action="{{ url_for('new_hassle_participants_submit') }}" method="POST">
+<form action="{{ url_for('hassle_bp.new_hassle_participants_submit') }}" method="POST">
   <table>
     <thead>
       <tr>
